requirements.txt
vosk==0.3.45
pyannote.audio==3.3.2
requests
flask